+++ optimism/client.go
@@ -21,7 +21,6 @@ import (
 	"log"
 	"math/big"
 	"net/http"
-	"reflect"
 	"strings"
 	"time"
 
@@ -34,7 +33,6 @@ import (
 	"github.com/ethereum-optimism/optimism/l2geth/params"
 	"github.com/ethereum-optimism/optimism/l2geth/rlp"
 	"github.com/ethereum-optimism/optimism/l2geth/rpc"
-	"github.com/inphi/optimism-rosetta/optimism/utilities/artifacts"
 
 	"github.com/ethereum/go-ethereum/crypto"
 	"golang.org/x/sync/semaphore"
@@ -996,145 +994,6 @@ func decodeHexData(data string) (*big.Int, error) {
 		return nil, fmt.Errorf("could not extract data from %s", data)
 	}
 	return decoded, nil
-}
-
-// Balance returns the balance of a *RosettaTypes.AccountIdentifier
-// at a *RosettaTypes.PartialBlockIdentifier.
-// The OP Token and ETH balances will be returned if currencies is unspecified
-//
-//nolint:gocognit
-func (ec *Client) Balance(
-	ctx context.Context,
-	account *RosettaTypes.AccountIdentifier,
-	block *RosettaTypes.PartialBlockIdentifier,
-	currencies []*RosettaTypes.Currency,
-) (*RosettaTypes.AccountBalanceResponse, error) {
-	var raw json.RawMessage
-	if block != nil {
-		if block.Hash != nil {
-			if err := ec.c.CallContext(ctx, &raw, "eth_getBlockByHash", block.Hash, false); err != nil {
-				return nil, err
-			}
-		}
-		if block.Hash == nil && block.Index != nil {
-			if err := ec.c.CallContext(
-				ctx,
-				&raw,
-				"eth_getBlockByNumber",
-				hexutil.EncodeUint64(uint64(*block.Index)),
-				false,
-			); err != nil {
-				return nil, err
-			}
-		}
-	} else {
-		if err := ec.c.CallContext(ctx, &raw, "eth_getBlockByNumber", toBlockNumArg(nil), false); err != nil {
-			return nil, err
-		}
-	}
-	if len(raw) == 0 {
-		return nil, ethereum.NotFound
-	}
-
-	var head *types.Header
-	if err := json.Unmarshal(raw, &head); err != nil {
-		return nil, err
-	}
-
-	var (
-		balance hexutil.Big
-		nonce   hexutil.Uint64
-		code    string
-	)
-
-	blockNum := hexutil.EncodeUint64(head.Number.Uint64())
-	reqs := []rpc.BatchElem{
-		{Method: "eth_getBalance", Args: []interface{}{account.Address, blockNum}, Result: &balance},
-		{Method: "eth_getTransactionCount", Args: []interface{}{account.Address, blockNum}, Result: &nonce},
-		{Method: "eth_getCode", Args: []interface{}{account.Address, blockNum}, Result: &code},
-	}
-	if err := ec.c.BatchCallContext(ctx, reqs); err != nil {
-		return nil, err
-	}
-	for i := range reqs {
-		if reqs[i].Error != nil {
-			return nil, reqs[i].Error
-		}
-	}
-
-	nativeBalance := &RosettaTypes.Amount{
-		Value:    balance.ToInt().String(),
-		Currency: Currency,
-	}
-
-	var balances []*RosettaTypes.Amount
-	for _, curr := range currencies {
-		if reflect.DeepEqual(curr, Currency) {
-			balances = append(balances, nativeBalance)
-			continue
-		}
-
-		contractAddress := fmt.Sprintf("%s", curr.Metadata[ContractAddressKey])
-		_, ok := ChecksumAddress(contractAddress)
-		if !ok {
-			return nil, fmt.Errorf("invalid contract address %s", contractAddress)
-		}
-
-		balance, err := ec.getBalance(ctx, account.Address, blockNum, contractAddress)
-		if err != nil {
-			return nil, fmt.Errorf("err encountered for currency %s, token address %s; %v", curr.Symbol, contractAddress, err)
-		}
-		balances = append(balances, &RosettaTypes.Amount{
-			Value:    balance,
-			Currency: curr,
-		})
-	}
-
-	if len(currencies) == 0 {
-		opTokenBalance, err := ec.getBalance(ctx, account.Address, blockNum, opTokenContractAddress.String())
-		if err != nil {
-			return nil, fmt.Errorf("err getting OP token balance; %v", err)
-		}
-		balances = append(balances, nativeBalance, &RosettaTypes.Amount{
-			Value:    opTokenBalance,
-			Currency: OPTokenCurrency,
-		})
-	}
-
-	return &RosettaTypes.AccountBalanceResponse{
-		Balances: balances,
-		BlockIdentifier: &RosettaTypes.BlockIdentifier{
-			Hash:  head.Hash().Hex(),
-			Index: head.Number.Int64(),
-		},
-		Metadata: map[string]interface{}{
-			"nonce": int64(nonce),
-			"code":  code,
-		},
-	}, nil
-}
-
-func (ec *Client) getBalance(ctx context.Context, accountAddress string, blockNum string, contractAddress string) (string, error) {
-	erc20Data, err := artifacts.ERC20ABI.Pack("balanceOf", common.HexToAddress(accountAddress))
-	if err != nil {
-		return "", err
-	}
-	encodedERC20Data := hexutil.Encode(erc20Data)
-
-	callParams := map[string]string{
-		"to":   contractAddress,
-		"data": encodedERC20Data,
-	}
-	var resp string
-	if err := ec.c.CallContext(ctx, &resp, "eth_call", callParams, blockNum); err != nil {
-		return "", err
-	}
-	balance, err := decodeHexData(resp)
-	if err != nil {
-		return "", err
-	}
-
-	return balance.String(), nil
 }
 
 // GetBlockByNumberInput is the input to the call
